🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
https://linear.app/skyvern/issue/SKY-5868/max-recursion-error-in-app
<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Enhance workflow editor and debugger with new `Workspace` component, improved `FloatingWindow` interactions, and dynamic UI element management using `useRanker`.
>
>   - **Components**:
>     - Introduce `Workspace` component to manage workflow nodes, browser sessions, and UI interactions.
>     - Update `FloatingWindow` with `onFocus` and `onBlur` event handlers.
>   - **Workflow Management**:
>     - Refactor `FlowRenderer` to delegate node and edge management to `Workspace`.
>     - Add `useRanker` hook for dynamic z-index management of UI elements.
>   - **UI Enhancements**:
>     - Improve `WorkflowHeader` and `NodeHeader` with additional functionalities for running and saving workflows.
>     - Enhance `WorkflowNodeLibraryPanel` and `WorkflowParametersPanel` with better user interaction handling.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for f1c85c5fbd095502e813984048c4928e596dfc7b. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
